requirements.txt
fastapi==0.112.1
aiohttp==3.10.5
uvicorn==0.30.6
tortoise-orm==0.21.6
asyncpg==0.29.0
pydantic-settings==2.4.0
aerich==0.7.2
aiokafka==0.11.0
pytest==8.3.2
pytest-mock==3.14.0
coverage==7.6.1
pytest-asyncio==0.24.0